refactor: 메인 컴포넌트화
